# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SWAIN River Pollutant Visualizer Backend API",
    "description": "This is a backend API for the SWAIN River Pollutant Visualizer."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "device": {
      "get": {
        "summary": "Get all the device coordinates",
        "description": "Get all the device coordinates",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DeviceGet"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      },
      "post": {
        "summary": "Add a new device using longitude and latitude",
        "description": "Add a new device using longitude and latitude",
        "parameters": [
          {
            "name": "device",
            "in": "body",
            "description": "Device to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RiverPoint"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a device using longitude and latitude",
        "description": "Delete a device using longitude and latitude",
        "parameters": [
          {
            "name": "device",
            "in": "body",
            "description": "Device to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RiverPoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "device/{pointId}": {
      "post": {
        "summary": "Update a device using the point ID",
        "description": "Update a device using the point ID",
        "parameters": [
          {
            "name": "pointId",
            "in": "path",
            "description": "Point ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a device using the point ID",
        "description": "Delete a device using the point ID",
        "parameters": [
          {
            "name": "pointId",
            "in": "path",
            "description": "Point ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "RiverPoint": {
      "type": "object",
      "properties": {
        "lat": {
          "type": "number",
          "format": "double"
        },
        "lng": {
          "type": "number",
          "format": "double"
        }
      },
      "required": [
        "lat",
        "lng"
      ]
    },
    "ErrorModel": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "DeviceGet": {
      "type": "array",
      "items": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "orderInBranch": {
            "type": "number"
          },
          "branchId": {
            "type": "number"
          },
          "device": {
            "type": "boolean"
          }
        }
      }
    },
...
  }
}